requirements.txt
spotipy==2.20.0
python-dotenv==0.21.0
pillow==9.2.0
html2image==2.0.1
